LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/marker-animate-expected.txt


PASS 'color' animation
FAIL 'opacity' animation assert_equals: opacity expected "1" but got "0.5"
FAIL 'color' + 'opacity' animation assert_equals: opacity expected "1" but got "0.5"
list item
